// src/libs/Resource.d.ts
export declare namespace resource {
    interface typeGetresourcesClassify {
        code: number;
        data: Data;
        message: string;
    }
    export interface typeGetResourcesz {
        code:    number;
        data:    Datas;
        message: string;
    }
  }

export interface Data {
    data: Datum[];
}

export interface Datum {
    create_time: string;
    desc:        string;
    id:          string;
    img_url:     string;
    title:       string;
    update_time: string;
    user_id:     string;
}
// Generated by https://quicktype.io


export interface Datas {
    data:     Datums[];
    has_next: boolean;
    has_prev: boolean;
    total:    number;
}

export interface Datums {
    create_time:              string;
    desc:                     string;
    id:                       string;
    img_url:                  string;
    is_show:                  number;
    link:                     string;
    resources_classify_id:    string;
    resources_classify_title: string;
    sort:                     number;
    title:                    string;
    update_time:              string;
    user_id:                  string;
}
// Generated by https://quicktype.io